Tshark : update writer interface table correctly
pcap-ng trace can have interface description block in the middle of
trace.

Tshark does not check if a new interface has been found between two
calls to read_dump. As a result, if tshark tries to write to a file the
packet that it reads, it might not have the information about the
interface id referenced in the packet that it just read.

This patch detects that a new interface has been found by comparing the
size of the interfaces table of the reader and the writer after each
read. If the reader's table's size is smaller than the writer's table's
size then we copy the missing interfaces to the writer.

Change-Id: Ib684435d9e740500b43fa0c7b1fc0752a77b704d
Signed-off-by: Benjamin Hesmans <benjamin.hesmans@uclouvain.be>